Rename auth client service
