+++ packages/react-app/src/components/AppBar.tsx
@@ -16,7 +16,7 @@ import { googleLogout } from '@react-oauth/google';
 import logo from '../assets/Logo.svg';
 import { UserInfo } from '../lib/types';
 import { UserContext } from '../routes/AuthRoute';
-import AuthClient from '../services/googleauthclient';
+import AuthClient from '../services/authclient';
 
 function ResponsiveAppBar() {
   const [anchorElUser, setAnchorElUser] = useState<null | HTMLElement>(null);
